Move observer to his own dedicated package
While adding package documentation, it became clear
that the observer does not belong to controller.
